deleted tasks initial implementation

+++ tasked/Repository/TaskRepository.swift
@@ -14,6 +14,7 @@ class TaskRepository: ObservableObject {
     let db = Firestore.firestore()
     
     @Published var tasks = [CustomTask]()
+    @Published var deletedTasks = [CustomTask]()
     
     init(){
         let settings = FirestoreSettings()
@@ -26,9 +27,17 @@ class TaskRepository: ObservableObject {
         db.collection("tasks").addSnapshotListener { (querySnapshot, error) in
             if let querySnapshot = querySnapshot {
                 DispatchQueue.main.async {
-                    self.tasks = querySnapshot.documents.map { d in
-                        return CustomTask(id: d.documentID, title: d["title"] as? String ?? "", completed: d["completed"] as? Bool ?? false, owner: d["owner"] as? String ?? "", taskMembers: d["taskMembers"] as? [String] ?? [])
+                    for d in querySnapshot.documents {
+                        let task = CustomTask(id: d.documentID, title: d["title"] as? String ?? "", completed: d["completed"] as? Bool ?? false, deleted: d["deleted"] as? Bool ?? false, owner: d["owner"] as? String ?? "", taskMembers: d["taskMembers"] as? [String] ?? [])
+                        if task.deleted {
+                            self.deletedTasks.append(task)
+                        } else {
+                            self.tasks.append(task)
+                        }
                     }
+//                    self.tasks = querySnapshot.documents.map { d in
+//                        return CustomTask(id: d.documentID, title: d["title"] as? String ?? "", completed: d["completed"] as? Bool ?? false, deleted: d["deleted"] as? Bool ?? false, owner: d["owner"] as? String ?? "", taskMembers: d["taskMembers"] as? [String] ?? [])
+//                    }
                 }
             }
         }
@@ -36,7 +45,7 @@ class TaskRepository: ObservableObject {
     
     func addTask(_ task: CustomTask) {
         do {
-            let _ = try db.collection("tasks").addDocument(data: ["title" : task.title, "completed" : task.completed])
+            let _ = try db.collection("tasks").addDocument(data: ["title" : task.title, "completed" : task.completed, "taskMembers" : task.taskMembers, "deleted" : task.deleted])
         }
         catch {
             fatalError("Unable to encode task: \(error.localizedDescription)")
@@ -45,7 +54,7 @@ class TaskRepository: ObservableObject {
     
     func updateTask(_ task: CustomTask) {
         do {
-            try db.collection("tasks").document(task.id).setData(["title" : task.title, "completed" : task.completed])
+            try db.collection("tasks").document(task.id).setData(["title" : task.title, "completed" : task.completed, "taskMembers" : task.taskMembers, "deleted" : task.deleted])
         }
         catch {
             fatalError("Unable to encode task: \(error.localizedDescription)")
@@ -53,8 +62,26 @@ class TaskRepository: ObservableObject {
     }
     
     func removeTask(_ task: CustomTask) {
+        if task.deleted {
+            do {
+                try db.collection("tasks").document(task.id).delete()
+            }
+            catch {
+                fatalError("Unable to encode task: \(error.localizedDescription)")
+            }
+        } else {
+            do {
+                try db.collection("tasks").document(task.id).setData(["title" : task.title, "completed" : task.completed, "taskMembers" : task.taskMembers, "deleted" : true])
+            }
+            catch {
+                fatalError("Unable to encode task: \(error.localizedDescription)")
+            }
+        }
+    }
+    
+    func restoreTask(_ task: CustomTask) {
         do {
-            try db.collection("tasks").document(task.id).delete()
+            try db.collection("tasks").document(task.id).setData(["title" : task.title, "completed" : task.completed, "taskMembers" : task.taskMembers, "deleted" : false])
         }
         catch {
             fatalError("Unable to encode task: \(error.localizedDescription)")
